# openapi.yaml
openapi: 3.1.0
info:
  title: AV Suite Backend Foundation
  description: Rehabilitation exercises management backend platform
  version: 0.1.0
paths:
...
  /api/v1/patients/{patient_id}/documents:
    post:
      tags:
      - Documents
      summary: Upload Patient Document
      description: Register document metadata for a specific patient.
      operationId: upload_patient_document_api_v1_patients__patient_id__documents_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Patient Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientDocumentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Documents
      summary: List Patient Documents By Patient
      description: List documents for a specific patient with optional filtering.
      operationId: list_patient_documents_by_patient_api_v1_patients__patient_id__documents_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: patient_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Patient Id
      - name: treatment_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Treatment Id
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/DocumentCategory'
          - type: 'null'
          title: Category
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents:
    post:
      tags:
      - Documents
      summary: Create Document
      description: Register document metadata.
      operationId: create_document_api_v1_documents_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientDocumentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Documents
      summary: List Documents
      description: List documents for the authenticated clinic with optional filtering.
      operationId: list_documents_api_v1_documents_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: patient_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Patient Id
      - name: treatment_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Treatment Id
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/DocumentCategory'
          - type: 'null'
          title: Category
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{id}:
    get:
      tags:
      - Documents
      summary: Get Document
      description: Retrieve document metadata by ID.
      operationId: get_document_api_v1_documents__id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Documents
      summary: Update Document
      description: Update document metadata.
      operationId: update_document_api_v1_documents__id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientDocumentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Documents
      summary: Delete Document
      description: Delete a patient document record.
      operationId: delete_document_api_v1_documents__id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{id}/download:
    get:
      tags:
      - Documents
      summary: Download Document
      description: Download or retrieve private document location (authenticated,
        clinic-scoped).
      operationId: download_document_api_v1_documents__id__download_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DocumentCategory:
      type: string
      enum:
      - medical_report
      - prescription
      - lab_result
      - consent_form
      - x_ray_scan
      - id_proof
      - other
      title: DocumentCategory
      description: Categories for patient documents and medical attachments.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PatientDocumentCreate:
      properties:
        label:
          type: string
          maxLength: 255
          minLength: 1
          title: Label
          description: Human readable title or document label.
        file_type:
          type: string
          maxLength: 100
          minLength: 1
          title: File Type
          description: MIME type or file extension.
        category:
          allOf:
          - $ref: '#/components/schemas/DocumentCategory'
          description: Document classification category.
          default: other
        notes:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Notes
          description: Optional notes or observations.
        patient_id:
          type: string
          format: uuid
          title: Patient Id
          description: Target patient UUID.
        treatment_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Treatment Id
          description: Optional linked treatment session UUID.
        uploaded_by:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Uploaded By
          description: Optional uploader staff UUID.
        file_url:
          type: string
          maxLength: 1024
          minLength: 1
          title: File Url
          description: Storage location URL or path.
        file_size:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: File Size
          description: File size in bytes.
      type: object
      required:
      - label
      - file_type
      - patient_id
      - file_url
      title: PatientDocumentCreate
      description: Schema for registering a new patient document.
    PatientDocumentListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PatientDocumentResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      type: object
      required:
      - items
      - total
      - offset
      - limit
      title: PatientDocumentListResponse
      description: Schema representing a paginated list of patient documents.
    PatientDocumentResponse:
      properties:
        label:
          type: string
          maxLength: 255
          minLength: 1
          title: Label
          description: Human readable title or document label.
        file_type:
          type: string
          maxLength: 100
          minLength: 1
          title: File Type
          description: MIME type or file extension.
        category:
          allOf:
          - $ref: '#/components/schemas/DocumentCategory'
          description: Document classification category.
          default: other
        notes:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Notes
          description: Optional notes or observations.
        id:
          type: string
          format: uuid
          title: Id
        clinic_id:
          type: string
          format: uuid
          title: Clinic Id
        patient_id:
          type: string
          format: uuid
          title: Patient Id
        treatment_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Treatment Id
        uploaded_by:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Uploaded By
        file_url:
          type: string
          title: File Url
        file_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: File Size
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - label
      - file_type
      - id
      - clinic_id
      - patient_id
      - treatment_id
      - uploaded_by
      - file_url
      - file_size
      - created_at
      - updated_at
      title: PatientDocumentResponse
      description: Schema representing a full patient document record.
    PatientDocumentUpdate:
      properties:
        label:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Label
        category:
          anyOf:
          - $ref: '#/components/schemas/DocumentCategory'
          - type: 'null'
        notes:
          anyOf:
          - type: string
            maxLength: 2000
          - type: 'null'
          title: Notes
      type: object
      title: PatientDocumentUpdate
      description: Schema for updating an existing document metadata record.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login